smart-energy-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
python-dotenv
groq
scikit-learn
numpy
joblib

